Let me explain what I've changed:
Instead of trying to handle the EAN code separately in the importer, I've moved it to the meta_data array in the mapper's prepare_meta_data method:
if (isset($baserow_data['EAN Code']) && !empty($baserow_data['EAN Code'])) {
    $ean = $this->sanitize_text_field($baserow_data['EAN Code']);
    $meta_data['_wpm_ean'] = $ean;
    $meta_data['_alg_ean'] = $ean;
    $meta_data['EAN'] = $ean;

    $this->log_debug("Added EAN code to meta data", [
        'ean' => $ean,
        'meta_keys' => ['_wpm_ean', '_alg_ean', 'EAN']
    ]);
}
Added logging to see if the EAN code is being added to the meta data

Added meta_data logging in the map_to_woocommerce method to verify it's being included in the final product data

This way, the EAN code will be handled along with all other meta data in one place, which should be more reliable than trying to set it separately.

Let me know if you want me to try a different approach.